backend/app/services/blockchain.py
"""
Blockchain Service for storing evidence on Polygon network.
Handles smart contract interactions and transaction management.
"""

from typing import Dict, Any
import json
from web3 import Web3
from eth_account import Account
from app.core.config import settings


class BlockchainService:
    """Service for interacting with Polygon blockchain"""
    
    def __init__(self):
        """Initialize blockchain service with Web3 connection"""
        try:
            self.w3 = Web3(Web3.HTTPProvider(settings.polygon_rpc_url))
            self.is_connected = self.w3.is_connected()
            self.contract_address = settings.contract_address
            self.private_key = settings.private_key
            
            # Load contract ABI
            try:
                with open("app/data/abi/EvidenceVault.json", "r") as f:
                    self.abi = json.load(f)
            except FileNotFoundError:
                self.abi = None
                
        except Exception as e:
            raise Exception(f"Failed to initialize blockchain service: {str(e)}")
    
    def store_evidence(self, evidence_data: str) -> Dict[str, Any]:
        """
        Store evidence on blockchain.
        
        Args:
            evidence_data: JSON string containing evidence data
            
        Returns:
            Dictionary containing transaction hash and confirmation details
        """
        try:
            # TODO: Implement contract interaction
            # Load contract ABI and interact with smart contract
            if not self.is_connected:
                return {
                    "status": "error",
                    "message": "Web3 not connected to blockchain",
                    "tx_hash": None
                }
            
            return {
                "status": "success",
                "tx_hash": "0x" + "0" * 64,
                "block_number": None,
                "message": "Evidence stored on blockchain (placeholder)"
            }
        except Exception as e:
            return {
                "status": "error",
                "message": str(e),
                "tx_hash": None
            }
    
    def retrieve_evidence(self, tx_hash: str) -> Dict[str, Any]:
        """
        Retrieve stored evidence from blockchain.
        
        Args:
            tx_hash: Transaction hash of stored evidence
            
        Returns:
            Dictionary containing retrieved evidence
        """
        try:
            if not self.is_connected:
                return {
                    "status": "error",
                    "message": "Web3 not connected to blockchain"
                }
            
            # TODO: Implement contract read operation
            return {
                "status": "success",
                "evidence": {},
                "verified": True,
                "timestamp": None
            }
        except Exception as e:
            return {
                "status": "error",
                "message": str(e)
            }
    
    def verify_evidence_integrity(self, evidence_hash: str) -> Dict[str, Any]:
        """
        Verify the integrity of stored evidence.
        
        Args:
            evidence_hash: Hash of the evidence to verify
            
        Returns:
            Dictionary containing verification results
        """
        try:
            if not self.is_connected:
                return {
                    "status": "error",
                    "message": "Web3 not connected to blockchain"
                }
            
            # TODO: Implement evidence verification
            return {
                "status": "success",
                "verified": True,
                "timestamp": None
            }
        except Exception as e:
            return {
                "status": "error",
                "message": str(e)
            }


def store_evidence(evidence_data: str) -> str:
    """
    Store evidence on blockchain.
    
    Args:
        evidence_data: JSON string containing evidence data
        
    Returns:
        Transaction hash
    """
    try:
        service = BlockchainService()
        result = service.store_evidence(evidence_data)
        return result.get("tx_hash", "")
    except Exception as e:
        # Return empty string on error for backward compatibility
        return ""


def retrieve_evidence(tx_hash: str) -> Dict[str, Any]:
    """
    Retrieve stored evidence from blockchain.
    
    Args:
        tx_hash: Transaction hash
        
    Returns:
        Evidence data
    """
    try:
        service = BlockchainService()
        return service.retrieve_evidence(tx_hash)
    except Exception as e:
        return {"status": "error", "message": str(e)}
